# agent/api/openapi.json
{
  "openapi":"3.0.3",
  "info": {
    "title":"QueryChat schema",
    "version":"0.0.1"
  },
  "servers": [
    {"url":"https://localhost:8080"}
  ],
  "tags": [
    {"name":"chat"}
  ],
  "paths": {
    "/v0.0.1/chat": {
      "post": {
        "tags": ["chat"],
        "summary":"Send message to the server",
        "operationId": "sendMSG",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/MSG"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description":"Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/VisualizableData"
                }
              }
            }
          },
          "default": {
            "description":"Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MSG": {
        "type":"object",
        "required": ["dbType, prompt"],
        "properties": {
          "dbType": {
            "type":"string"
          },
          "prompt": {
            "type":"string"
          }
        }
      },
      "VisualizableData": {
        "type":"object",
        "required": ["visualizableData"],
        "properties": {
          "visualizableData": {
            "type":"string"
          }
        }
      },
      "Error": {
        "description": "Represents error object",
        "type":"object",
        "properties": {
          "code": {
            "type":"integer",
            "format":"int64"
          },
          "message": {
            "type":"string"
          }
        },
        "required": ["code", "message"]
      }
    }
  }
}